+++ tapis-jobslib/src/main/java/edu/utexas/tacc/tapis/jobs/launchers/SingularityRunSlurmLauncher.java
@@ -2,10 +2,6 @@ package edu.utexas.tacc.tapis.jobs.launchers;
 
 import static edu.utexas.tacc.tapis.shared.utils.TapisUtils.alwaysSingleQuote;
 
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -30,10 +26,6 @@ public final class SingularityRunSlurmLauncher
     // Tracing.
     private static final Logger _log = LoggerFactory.getLogger(SingularityRunSlurmLauncher.class);
 
-    // Initialize the regex pattern that extracts the ID slurm assigned to the job.
-    // The regex ignores leading and trailing whitespace and groups the numeric ID.
-    private static final Pattern _resultPattern = Pattern.compile("\\s*Submitted batch job (\\d+)\\s*");
-    
     /* ********************************************************************** */
     /*                              Constructors                              */
     /* ********************************************************************** */
@@ -77,7 +69,7 @@ public final class SingularityRunSlurmLauncher
         
         // Let's see what happened.
         if (exitStatus != 0) {
-            String msg = MsgUtils.getMsg("JOBS_SUBMIT_ERROR", getClass().getSimpleName(), 
+            String msg = MsgUtils.getMsg("JOBS_SUBMIT_ERROR2", getClass().getSimpleName(),
                                          _job.getUuid(), cmd, result, exitStatus);
             throw new JobException(msg);
         }
@@ -91,7 +83,7 @@ public final class SingularityRunSlurmLauncher
         
         // -------------------- Get ID ------------------------------
         // Extract the slurm id.
-        String id = getSlurmId(result, cmd);
+        String id = JobUtils.getSlurmId(_job, result, cmd);
         
         // Save the id.
         _jobCtx.getJobsDao().setRemoteJobId(_job, id);
@@ -112,43 +104,5 @@ public final class SingularityRunSlurmLauncher
         // as an absolute path on the system.
         String execDir = JobExecutionUtils.getExecDir(_jobCtx, _job);
         return String.format("cd %s;sbatch %s", alwaysSingleQuote(execDir), JobExecutionUtils.JOB_WRAPPER_SCRIPT);
-    }
-
-    /* ********************************************************************** */
-    /*                            Private Methods                             */
-    /* ********************************************************************** */
-    /* ---------------------------------------------------------------------- */
-    /* getSlurmId:                                                            */
-    /* ---------------------------------------------------------------------- */
-    private String getSlurmId(String output, String cmd) throws JobException
-    {
-        // We have a problem if the result is not the slurm id.
-        if (StringUtils.isBlank(output)) {
-            String msg = MsgUtils.getMsg("JOBS_SLURM_SBATCH_NO_RESULT",  
-                                         _job.getUuid(), cmd);
-            throw new JobException(msg);
-        }
-        
-        // Strip any banner information from the remote result.
-        output = JobUtils.getLastLine(output.strip());
-        
-        // Look for the success message
-        Matcher m = _resultPattern.matcher(output);
-        var found = m.matches();
-        if (!found) {
-            String msg = MsgUtils.getMsg("JOBS_SLURM_SBATCH_INVALID_RESULT",  
-                                         _job.getUuid(), output);
-            throw new JobException(msg);
-        }
-        
-        int groupCount = m.groupCount();
-        if (groupCount < 1) {
-            String msg = MsgUtils.getMsg("JOBS_SLURM_SBATCH_INVALID_RESULT",  
-                                         _job.getUuid(), output);
-            throw new JobException(msg);
-        }
-        
-        // Group 1 contains the slurm ID.
-        return m.group(1);
     }
 }