+++ tapis-jobsapi/CHANGELOG.md
@@ -9,6 +9,16 @@ You may also reference live-docs based on the openapi specification here:
 https://tapis-project.github.io/live-docs
 
 -----------------------
+## 1.7.1 - 2024-10-15
+
+### New Features:
+1. Improved Slurm response parsing.
+2. For cancel of a FORK type job, use pkill -P to include sub-processes. Use kill as backup command.
+
+### Bug fixes:
+1. Debug message fix for ZIP cancel command.
+
+-----------------------
 ## 1.7.0 - 2024-09-17
 
 ### Bug fixes: